Conflicts:
	hc/src/hc_amdgcn.ll

+++ ocml/src/trigredH.h
@@ -5,7 +5,7 @@
  * License. See LICENSE.TXT for details.
  *===------------------------------------------------------------------------*/
 
-extern int MATH_PRIVATE(trigred)(__private half *r, half x);
+extern short MATH_PRIVATE(trigred)(__private half *r, half x);
 extern half MATH_PRIVATE(sincosred)(half x, __private half *cp);
-extern CONSTATTR half MATH_PRIVATE(tanred)(half x, int regn);
+extern CONSTATTR half MATH_PRIVATE(tanred)(half x, short regn);
